Added part of the ProjectFilesTreeTableModel, fixed a few quirks in the getSize(...) helper, started hacking together a relativeTime(...) helper.
git-svn-id: svn://svn.jakeapp.com/trunk@883 cfed719c-3465-4289-a5de-482d64f03f8e

+++ gui/src/main/java/com/jakeapp/gui/swing/JakeStatusBar.java
@@ -189,7 +189,7 @@ public class JakeStatusBar extends JakeGuiComponent implements ConnectionStatus 
         int projectFileCount = getCore().getProjectFileCout(getProject());
         String filesStr = getResourceMap().getString(projectFileCount == 1 ? "projectFile" : "projectFiles");
 
-        int projectSizeTotal = getCore().getProjectSizeTotal(getProject());
+        long projectSizeTotal = getCore().getProjectSizeTotal(getProject());
         String projectSize = JakeMainHelper.getSize(projectSizeTotal);
 
         // update project statistics
